+++ src/libromdata/RomDataFactory.cpp
@@ -252,6 +252,7 @@ const RomDataFactoryPrivate::RomDataFns RomDataFactoryPrivate::romDataFns_magic[
 
 	// Handhelds
 	GetRomDataFns_addr(DMG, ATTR_HAS_THUMBNAIL | ATTR_HAS_METADATA, 0x104, 0xCEED6666),
+	GetRomDataFns_addr(DMG, ATTR_HAS_THUMBNAIL | ATTR_HAS_METADATA, 0x304, 0xCEED6666),	// headered
 	GetRomDataFns_addr(GameBoyAdvance, ATTR_HAS_THUMBNAIL | ATTR_HAS_METADATA, 0x04, 0x24FFAE51),
 	GetRomDataFns_addr(Lynx, ATTR_NONE, 0, 'LYNX'),
 	GetRomDataFns_addr(NGPC, ATTR_HAS_METADATA, 12, ' SNK'),